refactor(BoardLayout): use BoardRoutes Outlets
