+++ src/pages/board/index.tsx
@@ -1,0 +1,7 @@
+import { BoardUI } from "./interface";
+
+function Board() {
+  return <BoardUI />;
+}
+
+export { Board };
