+++ src/pages/board/interface.tsx
@@ -1,0 +1,7 @@
+import { AppPage } from "@components/layout/AppPage";
+
+function BoardUI() {
+  return <AppPage />;
+}
+
+export { BoardUI };
