+++ src/pages/board/outlets/boardlayout/config/select.ts
@@ -1,0 +1,7 @@
+const filterOptions = [
+  { id: "1", label: "Opcion 1", disabled: false },
+  { id: "2", label: "Opcion 2", disabled: false },
+  { id: "3", label: "Opcion 3", disabled: false },
+];
+
+export { filterOptions };
