+++ src/pages/home/config/apps.config.tsx
@@ -45,10 +45,4 @@ const logoutConfig = {
   logoutTitle: "Cerrar sesión",
 };
 
-const filterOptions = [
-  { id: "1", label: "Opcion 1", disabled: false },
-  { id: "2", label: "Opcion 2", disabled: false },
-  { id: "3", label: "Opcion 3", disabled: false },
-];
-
-export { appsConfig, navigationConfig, logoutConfig, filterOptions };
+export { appsConfig, navigationConfig, logoutConfig };
