+++ src/routes/board.tsx
@@ -3,36 +3,16 @@ import { Route, Routes } from "react-router-dom";
 import { ItemNotFound } from "@components/layout/ItemNotFound";
 import { requestNotFoundConfig } from "@pages/request/config/requestNotFound.config";
 import { ErrorPage } from "@src/components/layout/ErrorPage";
+import { Board } from "@src/pages/board";
+import { BoardLayout } from "@pages/board/outlets/boardlayout";
 
 function BoardRoutes() {
   return (
     <Routes>
-      <Route
-        path="/"
-        element={
-          <ItemNotFound
-            image={requestNotFoundConfig.image}
-            title={requestNotFoundConfig.title}
-            description={requestNotFoundConfig.description}
-            buttonDescription={requestNotFoundConfig.buttonDescription}
-            route={requestNotFoundConfig.route}
-          />
-        }
-      >
+      <Route path="/" element={<Board />}>
+        <Route path="/" element={<BoardLayout />} />
         <Route
-          path="board"
-          element={
-            <ItemNotFound
-              image={requestNotFoundConfig.image}
-              title={requestNotFoundConfig.title}
-              description={requestNotFoundConfig.description}
-              buttonDescription={requestNotFoundConfig.buttonDescription}
-              route={requestNotFoundConfig.route}
-            />
-          }
-        />
-        <Route
-          path="board/peticion/:solicitud_id"
+          path="/peticion/:solicitud_id"
           element={
             <ItemNotFound
               image={requestNotFoundConfig.image}
